% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature.R
\name{Feature}
\alias{Feature}
\title{Feature Interface}
\description{
Feature Interface
}
\section{Constructor}{
 Feature$new()
}

\section{Methods}{

\code{$getUniqueId()} Get Feature ID

\code{$getMZ()} Get Mass to Charge(M/Z) value of feature.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Feature$new()}}
\item \href{#method-getUniqueId}{\code{Feature$getUniqueId()}}
\item \href{#method-getMZ}{\code{Feature$getMZ()}}
\item \href{#method-set_py_obj}{\code{Feature$set_py_obj()}}
\item \href{#method-get_py_obj}{\code{Feature$get_py_obj()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUniqueId"></a>}}
\if{latex}{\out{\hypertarget{method-getUniqueId}{}}}
\subsection{Method \code{getUniqueId()}}{
Get Feature ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$getUniqueId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
feature ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMZ"></a>}}
\if{latex}{\out{\hypertarget{method-getMZ}{}}}
\subsection{Method \code{getMZ()}}{
Get Mass to Charge(M/Z) value of feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$getMZ()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
M/Z value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-set_py_obj}{}}}
\subsection{Method \code{set_py_obj()}}{
Setter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$set_py_obj(Py_obj)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Sets the underlying python object of Feature.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-get_py_obj}{}}}
\subsection{Method \code{get_py_obj()}}{
Getter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$get_py_obj()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the underlying python object of Feature.
}

}
}
