% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureMap.R
\name{FeatureMap}
\alias{FeatureMap}
\title{FeatureMap Interface}
\description{
FeatureMap Interface
}
\section{Constructor}{
 FeatureMap$new()
}

\section{Methods}{

\code{$getFeature()} Get specific feature by index.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FeatureMap$new()}}
\item \href{#method-getFeature}{\code{FeatureMap$getFeature()}}
\item \href{#method-set_py_obj}{\code{FeatureMap$set_py_obj()}}
\item \href{#method-get_py_obj}{\code{FeatureMap$get_py_obj()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMap$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFeature"></a>}}
\if{latex}{\out{\hypertarget{method-getFeature}{}}}
\subsection{Method \code{getFeature()}}{
Get specfic feature by index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMap$getFeature(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{index of feature.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Feature object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-set_py_obj}{}}}
\subsection{Method \code{set_py_obj()}}{
Setter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMap$set_py_obj(Py_obj)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Sets the underlying python object of FeatureMap.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-get_py_obj}{}}}
\subsection{Method \code{get_py_obj()}}{
Getter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureMap$get_py_obj()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the underlying python object of FeatureMap.
}

}
}
