% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureXMLFile.R
\name{FeatureXMLFile}
\alias{FeatureXMLFile}
\title{FeatureXMLFile Interface}
\description{
FeatureXMLFile Interface
}
\section{Constructor}{
 FeatureXMLFile$new()
}

\section{Methods}{

\code{$load()} Load details from MzML file.

\code{$store()} Store details in MzML file.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load}{\code{FeatureXMLFile$load()}}
\item \href{#method-store}{\code{FeatureXMLFile$store()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
Load Features from FeatureXML file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureXMLFile$load(filename, fmap)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of file.}

\item{\code{fmap}}{FeatureMap object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store"></a>}}
\if{latex}{\out{\hypertarget{method-store}{}}}
\subsection{Method \code{store()}}{
Store Features into a FeatureXML file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureXMLFile$store(filename, fmap)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of file.}

\item{\code{fmap}}{FeatureMap object.}
}
\if{html}{\out{</div>}}
}
}
}
