% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdXMLFile.R
\name{IdXMLFile}
\alias{IdXMLFile}
\title{IdXMLFile Interface}
\description{
IdXMLFile Interface
}
\section{Constructor}{
 IdXMLFile$new()
}

\section{Methods}{

\code{$load()} Load details from IdXML file.

\code{$store()} Store details in IdXML file.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load}{\code{IdXMLFile$load()}}
\item \href{#method-store}{\code{IdXMLFile$store()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
Load details from IdXML file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdXMLFile$load(filename, protein_ids, peptide_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the idXML file.}

\item{\code{protein_ids}}{list of ProteinIdentification objects.}

\item{\code{peptide_ids}}{list of PeptideIdentification objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store"></a>}}
\if{latex}{\out{\hypertarget{method-store}{}}}
\subsection{Method \code{store()}}{
Store details in IdXML file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdXMLFile$store(filename, protein_ids, peptide_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the idXML file.}

\item{\code{protein_ids}}{list of ProteinIdentification objects.}

\item{\code{peptide_ids}}{list of PeptideIdentification objects.}
}
\if{html}{\out{</div>}}
}
}
}
