% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSExperiment.R
\name{MSExperiment}
\alias{MSExperiment}
\title{MSExperiment Interface}
\description{
MSExperiment Interface
}
\section{Constructor}{
 MSExperiment$new()
}

\section{Methods}{

\code{$getMSLevels()} Get Spectrum Levels.

\code{$getNrSpectra()} Get number of Spectrums.

\code{$getSpectra()} Get list of Spectrums.

\code{$getSpectrum()} Get a specific Spectrum by position.

\code{$setSpectra()} Set the spectrums for experiment.

\code{$size()} Get total number of experiments.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MSExperiment$new()}}
\item \href{#method-getMSLevels}{\code{MSExperiment$getMSLevels()}}
\item \href{#method-getNrSpectra}{\code{MSExperiment$getNrSpectra()}}
\item \href{#method-getSpectra}{\code{MSExperiment$getSpectra()}}
\item \href{#method-getSpectrum}{\code{MSExperiment$getSpectrum()}}
\item \href{#method-setSpectra}{\code{MSExperiment$setSpectra()}}
\item \href{#method-size}{\code{MSExperiment$size()}}
\item \href{#method-set_py_obj}{\code{MSExperiment$set_py_obj()}}
\item \href{#method-get_py_obj}{\code{MSExperiment$get_py_obj()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMSLevels"></a>}}
\if{latex}{\out{\hypertarget{method-getMSLevels}{}}}
\subsection{Method \code{getMSLevels()}}{
Get Spectrum Levels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$getMSLevels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of spectrum levels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNrSpectra"></a>}}
\if{latex}{\out{\hypertarget{method-getNrSpectra}{}}}
\subsection{Method \code{getNrSpectra()}}{
Get number of Spectrums
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$getNrSpectra()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
total number of spectrums.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSpectra"></a>}}
\if{latex}{\out{\hypertarget{method-getSpectra}{}}}
\subsection{Method \code{getSpectra()}}{
Get list of Spectrums.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$getSpectra()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of spectrums.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSpectrum"></a>}}
\if{latex}{\out{\hypertarget{method-getSpectrum}{}}}
\subsection{Method \code{getSpectrum()}}{
Get a specific Spectrum.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$getSpectrum(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{spectrum number.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
spectrum.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSpectra"></a>}}
\if{latex}{\out{\hypertarget{method-setSpectra}{}}}
\subsection{Method \code{setSpectra()}}{
Set the spectrums for experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$setSpectra(spectra)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spectra}}{list of MSSpectrum objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
Get total number of experiment entities.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
total number of experiments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-set_py_obj}{}}}
\subsection{Method \code{set_py_obj()}}{
Setter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$set_py_obj(Py_obj)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Sets the underlying python object of MSExperiment.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-get_py_obj}{}}}
\subsection{Method \code{get_py_obj()}}{
Getter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSExperiment$get_py_obj()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the underlying python object of MSExperiment.
}

}
}
