% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSSpectrum.R
\name{MSSpectrum}
\alias{MSSpectrum}
\title{MSSpectrum Interface}
\description{
MSSpectrum Interface
}
\section{Constructor}{
 MSSpectrum$new()
}

\section{Methods}{

\code{$getMSLevel()} Get Spectrum Level.

\code{$setMSLevel()} Set Spectrum Level.

\code{$getRT()} Get the retention time.

\code{$setRT()} Set the retention time.

\code{$get_peaks()} Get the peak values(m/z and intensity) as a list containing two lists, for m/z and intensity.

\code{$set_peaks()} Set the peak values(m/z and intensity) as a list containing two lists, for m/z and intensity.

\code{$size()} Get the count of peaks.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MSSpectrum$new()}}
\item \href{#method-getMSLevel}{\code{MSSpectrum$getMSLevel()}}
\item \href{#method-setMSLevel}{\code{MSSpectrum$setMSLevel()}}
\item \href{#method-getRT}{\code{MSSpectrum$getRT()}}
\item \href{#method-setRT}{\code{MSSpectrum$setRT()}}
\item \href{#method-get_peaks}{\code{MSSpectrum$get_peaks()}}
\item \href{#method-set_peaks}{\code{MSSpectrum$set_peaks()}}
\item \href{#method-size}{\code{MSSpectrum$size()}}
\item \href{#method-set_py_obj}{\code{MSSpectrum$set_py_obj()}}
\item \href{#method-get_py_obj}{\code{MSSpectrum$get_py_obj()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMSLevel"></a>}}
\if{latex}{\out{\hypertarget{method-getMSLevel}{}}}
\subsection{Method \code{getMSLevel()}}{
Get Spectrum Level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$getMSLevel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Spectrum Level.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMSLevel"></a>}}
\if{latex}{\out{\hypertarget{method-setMSLevel}{}}}
\subsection{Method \code{setMSLevel()}}{
Set Spectrum Level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$setMSLevel(level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{level number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRT"></a>}}
\if{latex}{\out{\hypertarget{method-getRT}{}}}
\subsection{Method \code{getRT()}}{
Get Retention Time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$getRT()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
retention time
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRT"></a>}}
\if{latex}{\out{\hypertarget{method-setRT}{}}}
\subsection{Method \code{setRT()}}{
Set Retention Time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$setRT(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{RT for the spectrum}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-get_peaks}{}}}
\subsection{Method \code{get_peaks()}}{
Get peak list from the Spectrum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$get_peaks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of two lists, for m/z and intensity, corresponding to peaks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-set_peaks}{}}}
\subsection{Method \code{set_peaks()}}{
Set peak list for the Spectrum
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$set_peaks(peaks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peaks}}{list with two lists, for m/z and intensity, corresponding to peaks.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
Get the count of peaks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
count of peaks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-set_py_obj}{}}}
\subsection{Method \code{set_py_obj()}}{
Setter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$set_py_obj(Py_obj)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Sets the underlying python object of MSSpectrum.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_py_obj"></a>}}
\if{latex}{\out{\hypertarget{method-get_py_obj}{}}}
\subsection{Method \code{get_py_obj()}}{
Getter method for wrapped object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MSSpectrum$get_py_obj()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the underlying python object of MSSpectrum.
}

}
}
