% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzMLFile.R
\name{MzMLFile}
\alias{MzMLFile}
\title{MzMLFile Interface}
\description{
MzMLFile Interface
}
\section{Constructor}{
 MzMLFile$new()
}

\section{Methods}{

\code{$load()} Load details from MzML file.

\code{$store()} Store details in MzML file.
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load}{\code{MzMLFile$load()}}
\item \href{#method-store}{\code{MzMLFile$store()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
Load details from MzML file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzMLFile$load(filename, exp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the MzML file.}

\item{\code{exp}}{MSExperiment object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-store"></a>}}
\if{latex}{\out{\hypertarget{method-store}{}}}
\subsection{Method \code{store()}}{
Store details in MzML file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzMLFile$store(filename, exp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the MzML file.}

\item{\code{exp}}{MSExperiment object.}
}
\if{html}{\out{</div>}}
}
}
}
